Remove households page v1
