+++ .env.development
@@ -19,5 +19,4 @@ SESSION_PASSWORD=thisispasswordandshouldbelongerthan32characters
 MONGODB_URL=mongodb://127.0.0.1:27017/zetkin
 
 FEAT_AREAS=*
-FEAT_HOUSEHOLDS2=*
 FEAT_PERSON_NOTES=*
