+++ src/features/canvass/components/LocationDialog/pages/LocationPage.tsx
@@ -19,15 +19,13 @@ import estimateVisitedHouseholds from 'features/canvass/utils/estimateVisitedHou
 import { Msg, useMessages } from 'core/i18n';
 import messageIds from 'features/canvass/l10n/messageIds';
 import useBasicLocationStats from 'features/canvass/hooks/useBasicLocationStats';
-import { HOUSEHOLDS2 } from 'utils/featureFlags';
-import useFeatureWithOrg from 'utils/featureFlags/useFeatureWithOrg';
 
 type LocationPageProps = {
   assignment: ZetkinAreaAssignment;
   location: ZetkinLocation;
   onClose: () => void;
   onEdit: () => void;
-  onHouseholds: (useNew: boolean) => void;
+  onHouseholds: () => void;
   onVisit: () => void;
 };
 
@@ -39,10 +37,6 @@ const LocationPage: FC<LocationPageProps> = ({
   onVisit,
   location,
 }) => {
-  const hasHouseholds2 = useFeatureWithOrg(
-    HOUSEHOLDS2,
-    assignment.organization_id
-  );
   const messages = useMessages(messageIds);
   const visits = uselocationVisits(
     assignment.organization_id,
@@ -92,14 +86,9 @@ const LocationPage: FC<LocationPageProps> = ({
             <Msg id={messageIds.location.page.quickReportButtonLabel} />
           </Button>
         )}
-        <Button onClick={() => onHouseholds(false)} variant="contained">
+        <Button onClick={() => onHouseholds()} variant="contained">
           <Msg id={messageIds.location.page.householdsButtonLabel} />
         </Button>
-        {hasHouseholds2 && (
-          <Button onClick={() => onHouseholds(true)} variant="contained">
-            Households v2
-          </Button>
-        )}
       </Box>
       <Box my={2}>
         <Divider />
