+++ src/utils/featureFlags/index.ts
@@ -4,6 +4,5 @@ export const AREAS = 'FEAT_AREAS';
 export const CALL = 'FEAT_CALL';
 export const OFFICIALS = 'FEAT_OFFICIALS';
 export const TASKS = 'FEAT_TASKS';
-export const HOUSEHOLDS2 = 'FEAT_HOUSEHOLDS2';
 export const PERSON_NOTES = 'FEAT_PERSON_NOTES';
 export const UNAUTH_EVENT_SIGNUP = 'FEAT_UNAUTH_EVENT_SIGNUP';
